merge development ready for 1.0.0 release
* Refactor: Mass clean

* HotFix: Issue with 'This' in ClashMode

* fix: ClashMode self and battle issue

Removed the use of self in the BattleEven method as I realised it wasn't a issue.
Fixed bug where at the start of the battlefield a odd and the adjacent even rows could
be the CellType.Defeat which caused visual issue when rendering and shouldn't happen anyway.

commit c0b4d8b7060a1c89f9e3bca57dca28b82b55b3ee

* refactor: ClashMode Even parameters name change

* feat: ClashMode ColourTwoMaxLenghtPercent

Change ColourTwoMaxLenghtPercent to allow a percent or a number <= BattleField Size

* feat: ClashMode paramter ColourOneMaxLenghtPercent added

This allows you to set the max lenght of the ColourOne line. Either with a number or a percent

* fix: ClashMode FinalStand incorrect opperactor

* feat: ClashMode Option to remove last victory

Added the option via configs to check the last battlefield for a single victory and remove it.
Reason for doing this was it looked very noticable and out of place.
Added a config to change the percent of it happening.

* fix: IMode parameter changed to be any

* fix: IRLG check for renderer config

* feat: Adds method to call next and draw together

This made one medthod which calls the Next() and Draw() method together.
Meaning you don't need to call both method however still can if need be.

BREAKING CHANGE:
This changes the Next() to Generate() and changes Next() to do both

* fix: Index missing a few exports

* refactor: Added error message for missing params in constructors

* docs: added JsDocs to everything

* feat: Css gradient to Canvas gradient

Adds functions for converting css gradients to canvas gradients

* refactor: Moving a utilities functions

Moved a utility function around and a few new basics ones

* feat: Canvas Renderer ability to use gradients

Can now pass in css gradients into as colour 1 and colour 2 as css gradient

* fix: css gradient angle calculation incorrect

Math is hard

* feat: Export on renderers

Added export to the Irenderer to allow the calling of Export().
This allows CanvasRenderer to export a image or return buffer for nodejs.

BREAKING CHANGE: IRenderer updated to have new method Export(opt:any)

* build: Changed script names

* build: Yarn to NPM

* test: Karma + Jasmine

Setup up karma with jasmine.
Added few test for testing purposes.

* test: Add tests for Utilities

* test: Add test for grid class

* test: Add ClashMode,BattleField and Grid tests

* fix: IsOdd non-number and floating

Check for non numbers and fix to work with floating point numbers

* test: fix IsOdd tests

* fix: IsOdd validation and test

* fix: IsHex validation

* test: Change to Karma config

* test: Update ClashMode test

* refactor: Add function to check if IsNumeric

* fix: ClashMode constructor validtion

Fix ClashMode constructor validation for SeedRand
Fix ClashMode tests

* fix: Css Gradient validation & test

* refactor: CanvasRenderer setting colours

* refactor: ClashMode percent moved percent calculation

* feat: Add ClashMode vertical

This took a copy and modified the ClashMode to work in vertical format
New mode added ClashModeVertical

* feat: Add virtical canvas Renderer

Added canvas renderer to draw verical

* feat: horizontal display and calcutions

Change virtical to horizontal, got this wrong way round.
Remove the copy and modifed ClashMode and CanvasRenderer virtical due finding smarter and
easier why to do this via rotating canvas and flipping height and width for grid.
Change between vertical and horizontal via a Config boolen.

BREAKING CHANGE: removal of ClashModeVertical and CanvasRendererVertical

* feat: add ScatterMode

Add the ScatterMode as a new mode/pattern

* feat: Add PincerMode

Add new mode/pattern

* refactor: ClashMode variable names

Change ClashMode variable names and the config names

BREAKING CHANGE: names for mode config changed

* fix: remove console logs in PincerMode

* refactor: ScatterMode variable names

Change ScatterMode variable names and the config names

BREAKING CHANGE: names for mode config changed

* refactor: PincerMode variable names

Change PincerMode variable names and the config names

BREAKING CHANGE: names for mode config changed

* build: change libray name

Update the libray to final name.
Update webpack config to align with name change.

BREAKING CHANGE: libray name change

* refactor: add method to draw graident

* fix: CanvasRenderer colour two not being exported

* fix: change gradient calculation

Change how the gradient was being calculated.
Was being calculated too well, was scaling and clipping gradient to make sure it fits
no matter what angle. This was causing issues for vertial gradients which I didn't like the
look of.
Both functions remain however old one isn't used but still exported.

BREAKING CHANGE: gradient calculated differently

* fix: node buffer change

New Buffer in node is deprecated, changed to use Buffer.from

* test: fix test for replaced function and config

* build: change webpack ts loader to use ts-loader

* fix: exports missing .js extension

es6 import for broswer not working

* build: webpack config to exclude node polyfill

* Create LICENSE

* fix: imports missing .js extension

* docs: add contributing.md

* docs: fix contributing.md

* Update CONTRIBUTING.md

Stupid line breaks

* Update CONTRIBUTING.md

Stupid markdown

* Update README.md

* fix: last row not calculated

Last row in the battlefield wasn't being calculated meaning if the rows were not an odd
number then it would not calculated it.

* fix: canvas renderer not draw on horizontal mode

The canvas renderer wasn't draw the right side when in horizontal mode

* ci: update package.json

* ci: update package.json

* Update package.json

* test: fixes broken test

This is due to a fix regarding the modes

d58d224199139702ab05ea6c67c3ddb77704ca30